requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
timm==0.9.9
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.22.0
thefuzz
requests==0.19.0
python-Levenshtein==0.21.1
PyMuPDF>=1.23.8
pydantic==2.5.0
python-dateutil==2.8.2
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
jinja2==3.1.2

